backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
motor==3.7.1
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.1
google-generativeai==0.4.0
email-validator==2.1.0
